Add an error manager in the back
