+++ package.json
@@ -68,6 +68,7 @@
   "devDependencies": {
     "@types/expect": "^24.3.0",
     "@types/express": "^4.17.11",
+    "@types/express-jwt": "^6.0.1",
     "@types/mocha": "^8.2.2",
     "@types/node": "^15.0.2",
     "@typescript-eslint/eslint-plugin": "^4.22.1",
